requirements.txt
pytest==7.2.0
requests==2.28.1
playwright==1.27.1
pytest-base-url==2.0.0
pytest-excel==1.5.0
et-xmlfile==1.1.0
pytest-playwright==0.3.0
python-dotenv~=0.21.0
openpyxl==3.0.10